app/views/chat_view.py
"""Chat view controller."""

from AppKit import NSView, NSViewController, NSMakeRect, NSColor, NSTextField
from Foundation import NSObject
import logging

logger = logging.getLogger(__name__)


class ChatViewController(NSViewController):
    """Controller for the chat tab."""
    
    def init(self):
        self = super(ChatViewController, self).init()
        if self is None:
            return None
        
        view_rect = NSMakeRect(0, 0, 800, 600)
        view = NSView.alloc().initWithFrame_(view_rect)
        view.setWantsLayer_(True)
        view.layer().setBackgroundColor_(NSColor.blackColor().CGColor())
        
        # Add chat placeholder
        chat_text = NSTextField.alloc().initWithFrame_(
            NSMakeRect(300, 300, 200, 50)
        )
        chat_text.setStringValue_("Chat with AI")
        chat_text.setEditable_(False)
        chat_text.setBordered_(False)
        chat_text.setBackgroundColor_(NSColor.clearColor())
        chat_text.setTextColor_(NSColor.lightGrayColor())
        
        view.addSubview_(chat_text)
        self.setView_(view)
        return self